move ShapeFuncInfo to its own file
